requirements.txt
streamlit>=1.28.0
google-genai>=0.3.0
python-dotenv>=1.0.0
playwright>=1.40.0
pytest>=7.4.0
pytest-playwright>=0.4.0